requirements.txt
py_clob_client==0.34.6
eth-account==0.13.7
web3==7.15.0
requests==2.33.1
python-dotenv==1.2.2
